refactor: extract mention syntax to core
